# fs: drop Stat() call from fsDeleteFile,deleteFile