## Description
This commit makes `fsDeleteFile()` simply call `deleteFile()` after calling
the relevant path length checking functions. This DRYs the code base.

This commit removes the `Stat()` call from `deleteFile()`. This improves
performance and removes any possibility of a race condition.

This additionally adds tests and a benchmark for said function. Here are
the results from before and after this commit:

```
benchmark                   old ns/op     new ns/op     delta
BenchmarkFSDeleteFile-4     2694884       1900661       -29.47%
```

<!--- Provide a general summary of your changes in the Title above -->


<!--- Describe your changes in detail -->

## Motivation and Context
Fixes `cmd/fs-v1-helpers.go` in https://github.com/minio/minio/issues/4658#issuecomment-318495587.
<!--- Why is this change required? What problem does it solve? -->
<!--- If it fixes an open issue, please link to the issue here. -->

## How Has This Been Tested?
Tests are added.
<!--- Please describe in detail how you tested your changes. -->
<!--- Include details of your testing environment, and the tests you ran to -->
<!--- see how your change affects other areas of the code, etc. -->

## Types of changes
<!--- What types of changes does your code introduce? Put an `x` in all the boxes that apply: -->
- [x] Bug fix (non-breaking change which fixes an issue)
- [ ] New feature (non-breaking change which adds functionality)
- [ ] Breaking change (fix or feature that would cause existing functionality to change)

## Checklist:
<!--- Go over all the following points, and put an `x` in all the boxes that apply. -->
<!--- If you're unsure about any of these, don't hesitate to ask. We're here to help! -->
- [x] My code follows the code style of this project.
- [ ] My change requires a change to the documentation.
- [ ] I have updated the documentation accordingly.
- [x] I have added tests to cover my changes.
- [x] All new and existing tests passed.